ark_bls12_377::g1

Type Alias G1TEAffine

pub type G1TEAffine = Affine<Config>;

Aliased Type

struct G1TEAffine {
    pub x: Fp<MontBackend<FqConfig, 6>, 6>,
    pub y: Fp<MontBackend<FqConfig, 6>, 6>,
}

Fields
x: Fp<MontBackend<FqConfig, 6>, 6>

X coordinate of the point represented as a field element

y: Fp<MontBackend<FqConfig, 6>, 6>

Y coordinate of the point represented as a field element