Crate ark_bls12_377

This library implements the BLS12_377 curve generated in [BCGMMW20, “Zexe”]. The name denotes that it is a Barreto–Lynn–Scott curve of embedding degree 12, defined over a 377-bit (prime) field. The main feature of this curve is that both the scalar field and the base field are highly 2-adic. (This is in contrast to the BLS12_381 curve for which only the scalar field is highly 2-adic.)

Curve information:

Base field: q = 258664426012969094010652733694893533536393512754914660539884262666720468348340822774968888139573360124440321458177

Scalar field: r = 8444461749428370424248824938781546531375899335154063827935233455917409239041

valuation(q - 1, 2) = 46

valuation(r - 1, 2) = 47

G1 curve equation: y^2 = x^3 + 1

G2 curve equation: y^2 = x^3 + B, where

B = Fq2(0, 155198655607781456406391640216936120121836107652948796323930557600032281009004493664981332883744016074664192874906)

Modules

fq

fq2

fq6

fq12

fr

g1

g2

Structs

Fq2Parameters

Fq6Parameters

Fq12Parameters

FqParameters

FrParameters

Parameters

Constants

FQ2_ONE

FQ2_ZERO

FQ_ONE

FQ_ZERO

Type Definitions

Bls12_377

Fq

Fq2

Fq6

Fq12

Fr

G1Affine

G1Projective

G2Affine

G2Projective